qt backend classes separated
